small entities refactoring

+++ src/main/java/com/rmr/dinosaurs/infrastucture/database/CourseAndProfessionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface CourseAndProfessionRepository
-    extends JpaRepository<CourseAndProfession, Integer> {
+    extends JpaRepository<CourseAndProfession, Long> {
 
 }
